Move update reconnections to connector class.
