+++ HNSWIndex/GraphConnector.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using System.Collections.Concurrent;
 using System.Numerics;
 
 namespace HNSWIndex
@@ -17,6 +18,10 @@ namespace HNSWIndex
             parameters = hnswParams;
         }
 
+        /// <summary>
+        /// Establish connections from newly inserted node to the graph.
+        /// If graph is empty, create new graph from a node.
+        /// </summary>
         internal void ConnectNewNode(int nodeId)
         {
             // If this is new ep we keep lock for entire Add Operation
@@ -42,6 +47,21 @@ namespace HNSWIndex
             }
         }
 
+        /// <summary>
+        /// Perform update of connections on provided list of indexes 
+        /// </summary>
+        internal void UpdateOldConnections(IList<int> indexes, IList<TLabel> labels)
+        {
+            if (indexes.Count != labels.Count) throw new ArgumentException("Update collections size mismatch");
+
+            var dirtyByIndex = DisconnectDirtyEdges(indexes, labels);
+            ReconnectDirtyNodes(dirtyByIndex);
+        }
+
+        /// <summary>
+        /// Remove node from the graph at all layers.
+        /// After this operation no other node in graph will point to provided item.
+        /// </summary>
         internal void RemoveNodeConnections(Node item)
         {
             // TODO: Think about different method for handling side effects which have to be done under neighborhood lock.
@@ -65,7 +85,10 @@ namespace HNSWIndex
             }
         }
 
-        internal void RemoveConnectionsAtLayer(Node removedNode, int layer)
+        /// <summary>
+        /// Remove node from the graph at given layer.
+        /// </summary>
+        private void RemoveConnectionsAtLayer(Node removedNode, int layer)
         {
             WipeRelationsWithNode(removedNode, layer);
 
@@ -101,9 +124,117 @@ namespace HNSWIndex
             }
         }
 
+        /// <summary>
+        /// Remove outgoing edge from node to invalid target.
+        /// This operation is assumed to work under neighborhood lock. 
+        /// </summary>
         private void RemoveOutEdge(Node target, Node badNeighbour, int layer)
         {
             target.OutEdges[layer].Remove(badNeighbour.Id);
+        }
+
+
+        /// <summary>
+        /// Mark edges with outdated neighborhood as dirty and disconnect them from the graph.
+        /// Return dictionary with ids of dirty nodes valued by ditry level.
+        /// </summary>
+        private ConcurrentDictionary<int, int> DisconnectDirtyEdges(IList<int> indexes, IList<TLabel> labels)
+        {
+            // index -> highest layer that needs reinsertion, -1 means clean
+            var dirtyByIndex = new ConcurrentDictionary<int, int>(Environment.ProcessorCount * 2, indexes.Count);
+            var cleanAnchors = new int[data.GetTopLayer() + 1];
+            Array.Fill(cleanAnchors, data.EntryPointId);
+
+            // Decide which layers to rewire and temporarily disconnect at those layers
+            Parallel.For(0, indexes.Count, (i) =>
+            {
+                var index = indexes[i];
+                var newLabel = labels[i];
+                var oldLabel = data.Items[index];
+                var difference = data.Distance(newLabel, oldLabel);
+                var node = data.Nodes[index];
+
+                if (TDistance.IsZero(difference)) return;
+
+                for (int layer = 0; layer <= node.MaxLayer; layer++)
+                {
+                    // Update on significant difference 
+                    using (data.GraphLocker.LockNodeNeighbourhood(node, layer))
+                    {
+                        var outs = node.OutEdges[layer];
+                        var isDirtyAlready = dirtyByIndex.ContainsKey(index);
+                        if (outs.Count == 0) return; // Most likely single point at layer
+
+                        // TODO: Replace linq call with manual min finding
+                        var minEdge = outs.Count > 0 ? outs.Min(neighbor => data.Distance(index, neighbor)) : TDistance.MaxValue;
+                        if (difference < minEdge) return; // Skip layer w/o significant change
+
+                        dirtyByIndex[index] = layer;
+
+                        // Move ep if change would invalid its status
+                        // TODO: Replace linq call with manual maxby finding
+                        if (index == cleanAnchors[layer])
+                        {
+                            cleanAnchors[layer] = outs.Count > 0 ? outs.MaxBy(id => data.Nodes[id].OutEdges.Count) : -1;
+                        }
+                        RemoveConnectionsAtLayer(node, layer);
+                        ResetNodeConnectionsAtLayer(node, layer);
+                    }
+                }
+                // swap the label
+                data.Items[index] = newLabel;
+            });
+
+            // Resotore good ep
+            if (dirtyByIndex.GetValueOrDefault(data.EntryPointId, -1) >= 0)
+            {
+                for (int layer = dirtyByIndex[data.EntryPointId]; layer >= 0; layer--)
+                {
+                    var restorationEntry = data.Nodes[cleanAnchors[layer]];
+                    ConnectAtLayer(data.EntryPoint, restorationEntry, layer);
+                }
+                dirtyByIndex[data.EntryPointId] = -1;
+            }
+
+            return dirtyByIndex;
+        }
+
+        /// <summary>
+        /// Reestablish connections from dirty nodes.
+        /// </summary>
+        private void ReconnectDirtyNodes(ConcurrentDictionary<int, int> dirtyByIndex)
+        {
+            Parallel.ForEach(dirtyByIndex, (kvp) =>
+            {
+                var index = kvp.Key;
+                var topLayer = kvp.Value;
+                var node = data.Nodes[index];
+                if (topLayer < 0) return;
+
+                // Perform reconnect
+                lock (node.OutEdgesLock)
+                {
+                    // Select best peer which has connections at rebuild layer
+                    Func<int, int, bool> EntryFilter = (cand, layer) => cand != index && dirtyByIndex.GetValueOrDefault(cand, -1) < layer;
+                    var bestPeer = navigator.FindEntryPoint(topLayer, data.Items[index], true, cand => EntryFilter(cand, topLayer));
+                    for (int layer = topLayer; layer >= 0; layer--)
+                    {
+                        if (!EntryFilter(bestPeer.Id, layer)) throw new InvalidOperationException("Dirty entry point while reconnecting");
+                        if (layer > 0)
+                        {
+                            var nextLayer = layer - 1;
+                            var nextEntryCandidateId = ConnectAtLayer(node, bestPeer, layer, cand => EntryFilter(cand, nextLayer));
+                            bestPeer = nextEntryCandidateId >= 0 ? data.Nodes[nextEntryCandidateId] : data.EntryPoint;
+                        }
+                        else ConnectAtLayer(node, bestPeer, layer);
+
+                        // update status
+                        var currDirtyLevel = dirtyByIndex.TryGetValue(index, out var a);
+                        dirtyByIndex.TryUpdate(index, a - 1, a);
+                    }
+                    dirtyByIndex.TryUpdate(index, -1, 0);
+                }
+            });
         }
 
         /// <summary>
@@ -176,6 +307,9 @@ namespace HNSWIndex
             node.OutEdges[layer] = newNeighbours;
         }
 
+        /// <summary>
+        /// Clear nodes adjacency list across all layers.
+        /// </summary>
         internal void ResetNodeConnections(Node node)
         {
             for (int layer = node.MaxLayer; layer >= 0; layer--)
@@ -184,12 +318,18 @@ namespace HNSWIndex
             }
         }
 
+        /// <summary>
+        /// Clear node adjacencty list at specified layer.
+        /// </summary>
         internal void ResetNodeConnectionsAtLayer(Node node, int layer)
         {
             node.OutEdges[layer] = new List<int>(data.MaxEdges(layer) + 1);
             node.InEdges[layer] = new List<int>(data.MaxEdges(layer) + 1);
         }
 
+        /// <summary>
+        /// Forget node as neighbor by incomming edge.
+        /// </summary>
         private void WipeRelationsWithNode(Node node, int layer)
         {
             lock (node.OutEdgesLock)
@@ -204,6 +344,9 @@ namespace HNSWIndex
             }
         }
 
+        /// <summary>
+        /// Set incoming edges to node based on its adjacency list.
+        /// </summary>
         private void SetRelationsWithNode(Node node, int layer)
         {
             lock (node.OutEdgesLock)
