+++ HNSWIndex/HNSWIndex.cs
@@ -107,96 +107,7 @@ namespace HNSWIndex
         /// </summary>
         public void Update(IList<int> indexes, IList<TLabel> labels)
         {
-            if (indexes.Count != labels.Count) throw new ArgumentException("Update collections size mismatch");
-
-            // index -> highest layer that needs reinsertion, -1 means clean
-            var dirtyByIndex = new ConcurrentDictionary<int, int>(Environment.ProcessorCount * 2, indexes.Count);
-            var cleanAnchors = new int[data.GetTopLayer() + 1];
-            Array.Fill(cleanAnchors, data.EntryPointId);
-
-            // Decide which layers to rewire and temporarily disconnect at those layers
-            Parallel.For(0, indexes.Count, (i) =>
-            {
-                var index = indexes[i];
-                var newLabel = labels[i];
-                var oldLabel = data.Items[index];
-                var difference = data.Distance(newLabel, oldLabel);
-                var node = data.Nodes[index];
-
-                if (TDistance.IsZero(difference)) return;
-
-                for (int layer = 0; layer <= node.MaxLayer; layer++)
-                {
-                    // Update on significant difference 
-                    using (data.GraphLocker.LockNodeNeighbourhood(node, layer))
-                    {
-                        var outs = node.OutEdges[layer];
-                        var isDirtyAlready = dirtyByIndex.ContainsKey(index);
-                        if (outs.Count == 0) return; // Most likely single point at layer
-
-                        // TODO: Replace linq call with manual min finding
-                        var minEdge = outs.Count > 0 ? outs.Min(neighbor => data.Distance(index, neighbor)) : TDistance.MaxValue;
-                        if (difference < minEdge) return; // Skip layer w/o significant change
-
-                        dirtyByIndex[index] = layer;
-
-                        // Move ep if change would invalid its status
-                        // TODO: Replace linq call with manual maxby finding
-                        if (index == cleanAnchors[layer])
-                        {
-                            cleanAnchors[layer] = outs.Count > 0 ? outs.MaxBy(id => data.Nodes[id].OutEdges.Count) : -1;
-                        }
-                        connector.RemoveConnectionsAtLayer(node, layer);
-                        connector.ResetNodeConnectionsAtLayer(node, layer);
-                    }
-                }
-                // swap the label
-                data.Items[index] = newLabel;
-            });
-
-            // Resotore good ep
-            if (dirtyByIndex.GetValueOrDefault(data.EntryPointId, -1) >= 0)
-            {
-                for (int layer = dirtyByIndex[data.EntryPointId]; layer >= 0; layer--)
-                {
-                    var restorationEntry = data.Nodes[cleanAnchors[layer]];
-                    connector.ConnectAtLayer(data.EntryPoint, restorationEntry, layer);
-                }
-                dirtyByIndex[data.EntryPointId] = -1;
-            }
-
-            // Reinsert dirty nodes
-            Parallel.ForEach(dirtyByIndex, (kvp) =>
-            {
-                var index = kvp.Key;
-                var topLayer = kvp.Value;
-                var node = data.Nodes[index];
-                if (topLayer < 0) return;
-
-                // Perform reconnect
-                lock (node.OutEdgesLock)
-                {
-                    // Select best peer which has connections at rebuild layer
-                    Func<int, int, bool> EntryFilter = (cand, layer) => cand != index && dirtyByIndex.GetValueOrDefault(cand, -1) < layer;
-                    var bestPeer = navigator.FindEntryPoint(topLayer, data.Items[index], true, cand => EntryFilter(cand, topLayer));
-                    for (int layer = topLayer; layer >= 0; layer--)
-                    {
-                        if (!EntryFilter(bestPeer.Id, layer)) throw new InvalidOperationException("Dirty entry point while reconnecting");
-                        if (layer > 0)
-                        {
-                            var nextLayer = layer - 1;
-                            var nextEntryCandidateId = connector.ConnectAtLayer(node, bestPeer, layer, cand => EntryFilter(cand, nextLayer));
-                            bestPeer = nextEntryCandidateId >= 0 ? data.Nodes[nextEntryCandidateId] : data.EntryPoint;
-                        }
-                        else connector.ConnectAtLayer(node, bestPeer, layer);
-
-                        // update status
-                        var currDirtyLevel = dirtyByIndex.TryGetValue(index, out var a);
-                        dirtyByIndex.TryUpdate(index, a - 1, a);
-                    }
-                    dirtyByIndex.TryUpdate(index, -1, 0);
-                }
-            });
+            connector.UpdateOldConnections(indexes, labels);
         }
 
         /// <summary>
